LibWeb/WebGL: Move extensions into their own folder
