+++ Libraries/LibWeb/CMakeLists.txt
@@ -823,9 +823,10 @@ set(SOURCES
     WebDriver/Response.cpp
     WebDriver/Screenshot.cpp
     WebDriver/TimeoutsConfiguration.cpp
-    WebGL/ANGLEInstancedArrays.cpp
+    WebGL/Extensions/ANGLEInstancedArrays.cpp
+    WebGL/Extensions/OESVertexArrayObject.cpp
+    WebGL/Extensions/WebGLVertexArrayObjectOES.cpp
     WebGL/EventNames.cpp
-    WebGL/OESVertexArrayObject.cpp
     WebGL/OpenGLContext.cpp
     WebGL/WebGL2RenderingContext.cpp
     WebGL/WebGLActiveInfo.cpp
@@ -846,7 +847,6 @@ set(SOURCES
     WebGL/WebGLTransformFeedback.cpp
     WebGL/WebGLUniformLocation.cpp
     WebGL/WebGLVertexArrayObject.cpp
-    WebGL/WebGLVertexArrayObjectOES.cpp
     WebIDL/AbstractOperations.cpp
     WebIDL/Buffers.cpp
     WebIDL/CallbackType.cpp
