+++ Libraries/LibWeb/WebGL/WebGLRenderingContext.h
@@ -81,8 +81,8 @@ private:
 
     // Extensions
     // "Multiple calls to getExtension with the same extension string, taking into account case-insensitive comparison, must return the same object as long as the extension is enabled."
-    GC::Ptr<ANGLEInstancedArrays> m_angle_instanced_arrays_extension;
-    GC::Ptr<OESVertexArrayObject> m_oes_vertex_array_object_extension;
+    GC::Ptr<Extensions::ANGLEInstancedArrays> m_angle_instanced_arrays_extension;
+    GC::Ptr<Extensions::OESVertexArrayObject> m_oes_vertex_array_object_extension;
 
     virtual void set_error(GLenum error) override;
 };
