+++ Libraries/LibWeb/idl_files.cmake
@@ -380,8 +380,9 @@ libweb_js_bindings(WebAudio/OscillatorNode)
 libweb_js_bindings(WebAudio/PannerNode)
 libweb_js_bindings(WebAudio/PeriodicWave)
 libweb_js_bindings(WebAudio/StereoPannerNode)
-libweb_js_bindings(WebGL/ANGLEInstancedArrays)
-libweb_js_bindings(WebGL/OESVertexArrayObject)
+libweb_js_bindings(WebGL/Extensions/ANGLEInstancedArrays)
+libweb_js_bindings(WebGL/Extensions/OESVertexArrayObject)
+libweb_js_bindings(WebGL/Extensions/WebGLVertexArrayObjectOES)
 libweb_js_bindings(WebGL/WebGL2RenderingContext)
 libweb_js_bindings(WebGL/WebGLActiveInfo)
 libweb_js_bindings(WebGL/WebGLBuffer)
@@ -400,7 +401,6 @@ libweb_js_bindings(WebGL/WebGLTexture)
 libweb_js_bindings(WebGL/WebGLTransformFeedback)
 libweb_js_bindings(WebGL/WebGLUniformLocation)
 libweb_js_bindings(WebGL/WebGLVertexArrayObject)
-libweb_js_bindings(WebGL/WebGLVertexArrayObjectOES)
 libweb_js_bindings(WebIDL/DOMException)
 libweb_js_bindings(WebSockets/WebSocket)
 libweb_js_bindings(WebVTT/VTTCue)
