+++ Meta/Lagom/Tools/CodeGenerators/LibWeb/BindingsGenerator/IDLGenerators.cpp
@@ -4440,6 +4440,7 @@ using namespace Web::UserTiming;
 using namespace Web::WebAssembly;
 using namespace Web::WebAudio;
 using namespace Web::WebGL;
+using namespace Web::WebGL::Extensions;
 using namespace Web::WebIDL;
 using namespace Web::WebVTT;
 using namespace Web::XHR;
